+++ server/src/index.js
@@ -6,6 +6,7 @@ import { startScheduler, stopScheduler } from './services/digestScheduler.js';
 import { startEmailPoller, stopEmailPoller } from './services/emailPoller.js';
 import { startRecurringScheduler, stopRecurringScheduler } from './services/recurringScheduler.js';
 import { startAlertsScheduler, stopAlertsScheduler } from './services/alertsScheduler.js';
+import { startGcalScheduler, stopGcalScheduler } from './services/gcalScheduler.js';
 
 runMigrations();
 bootstrapAdmin();
@@ -14,6 +15,7 @@ startScheduler();
 startEmailPoller();
 startRecurringScheduler();
 startAlertsScheduler();
+startGcalScheduler();
 
 const app = createApp();
 const server = app.listen(config.port, () => {
@@ -26,6 +28,7 @@ const shutdown = (sig) => {
   stopEmailPoller();
   stopRecurringScheduler();
   stopAlertsScheduler();
+  stopGcalScheduler();
   server.close(() => process.exit(0));
 };
 process.on('SIGINT', shutdown);